Add performance benchmark tests for Live API and Pose Detection
- Implemented LiveApiPerformanceTest to evaluate WebSocket connection time, voice response latency, audio streaming performance, network resilience, and memory usage under load.
- Added PoseDetectionPerformanceTest to assess pose detection latency, frame processing throughput, memory usage, CPU usage, and accuracy under performance constraints.
- Both tests follow a TDD approach, ensuring performance requirements are met and optimized for real-world scenarios.

+++ app/src/main/kotlin/com/posecoach/app/livecoach/audio/AudioQualityMonitor.kt
@@ -143,7 +143,7 @@ class AudioQualityMonitor(
     private fun calculateNoisePower(buffer: ShortArray): Double {
         // Estimate noise power from silent regions
         val threshold = SILENCE_THRESHOLD
-        val silentSamples = buffer.filter { abs(it) < threshold }
+        val silentSamples = buffer.filter { abs(it.toInt()) < threshold }
 
         return if (silentSamples.isNotEmpty()) {
             silentSamples.map { it.toDouble() * it.toDouble() }.average()
@@ -154,12 +154,12 @@ class AudioQualityMonitor(
 
     private fun detectClipping(buffer: ShortArray): Boolean {
         val clippingThreshold = (Short.MAX_VALUE * CLIPPING_THRESHOLD_RATIO).toInt()
-        return buffer.any { abs(it) >= clippingThreshold }
+        return buffer.any { abs(it.toInt()) >= clippingThreshold }
     }
 
     private fun calculateClippingPercentage(buffer: ShortArray): Double {
         val clippingThreshold = (Short.MAX_VALUE * CLIPPING_THRESHOLD_RATIO).toInt()
-        val clippedSamples = buffer.count { abs(it) >= clippingThreshold }
+        val clippedSamples = buffer.count { abs(it.toInt()) >= clippingThreshold }
         return clippedSamples.toDouble() / buffer.size
     }
